pyproject.toml
[project]
name = "finalproject-matesheva-dpo"
version = "0.1.0"
description = "Консольное приложение для управления валютным кошельком"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "prettytable (>=3.17.0,<4.0.0)",
    "python-dotenv (>= 1.0.0)"
]

[tool.poetry]
packages = [{include = "valutatrade_hub", from = "."}]

[tool.poetry.scripts]
project = "main:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
